feat(filter): blocksum・gamma_trc を C 準拠に、fill_map_holes に filltype (GREEN)
実装差異 #44: blocksum の正規化が 1 パスの f64 丸めだった。C blocksumLow は

1. 全カーネル面積の norm = 255/(fwc*fhc) で正規化し、f32 の積を byte に
   切り捨てる
2. 境界の行・列を、切り捨て済みの byte に対して fhc/hn・fwc/wn で
   再スケールし、また切り捨てる

という 2 パス構成。理想値を 1 回丸めるのとは多くの画素で 1 ずれる
(全 ON 画像でも角が 252 になる)。accumulator も C 同様 1bpp を
「ON 画素数」で積算するよう blockconv_accum を 1bpp 対応にした。

実装差異 #45: fill_map_holes が filltype を取らず L_FILL_BLACK 固定
だった。C は `valtest = (filltype == L_FILL_WHITE) ? 255 : 0` で穴の値を
切り替える。MapFillType を導入した。

実装差異 #46: gamma_trc の LUT が `255. * powf(x, invgamma) + 0.5` を
f32 で評価していた。C の `255.` は double リテラルのため倍精度評価に
なり、.5 境界に乗る値が 1 ずれる (maxval 270 で入力 153 が C 144 に対し
145)。

convolve の C check 2-4 と adaptmap の C check 14-15 を 5 ペアマップ
— 全件 Ok (Ok 233 → 238)。

+++ src/filter/enhance.rs
@@ -55,7 +55,12 @@ pub fn gamma_trc(gamma: f32, minval: i32, maxval: i32) -> FilterResult<TrcLut> {
             255
         } else {
             let x = (i - minval) as f32 / range;
-            let mapped = 255.0 * x.powf(inv_gamma) + 0.5;
+            // C: `val = (l_int32)(255. * powf(x, invgamma) + 0.5)`. The
+            // `255.` literal is a double, so the product and the rounding
+            // are evaluated in double precision even though `x` and `powf`
+            // are single. Doing it all in f32 shifts values that land near
+            // a .5 boundary by one (e.g. 153 with maxval 270: 144 vs 145).
+            let mapped = 255.0f64 * x.powf(inv_gamma) as f64 + 0.5;
             (mapped as i32).clamp(0, 255)
         };
         lut[i as usize] = val as u8;
